Remove a truck load of unused code
