+++ auth/src/main/java/com/firebase/ui/auth/ui/email/CheckEmailFragment.java
@@ -22,8 +22,13 @@ import com.firebase.ui.auth.ui.FlowParameters;
 import com.firebase.ui.auth.ui.TaskFailureLogger;
 import com.firebase.ui.auth.ui.User;
 import com.firebase.ui.auth.ui.email.fieldvalidators.EmailFieldValidator;
-import com.firebase.ui.auth.util.FirebaseAuthWrapperFactory;
+import com.firebase.ui.auth.util.GoogleApiConstants;
+import com.google.android.gms.auth.api.Auth;
 import com.google.android.gms.auth.api.credentials.Credential;
+import com.google.android.gms.auth.api.credentials.CredentialPickerConfig;
+import com.google.android.gms.auth.api.credentials.HintRequest;
+import com.google.android.gms.common.ConnectionResult;
+import com.google.android.gms.common.api.GoogleApiClient;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
@@ -216,16 +221,34 @@ public class CheckEmailFragment extends BaseFragment implements View.OnClickList
     }
 
     private void showEmailAutoCompleteHint() {
-        PendingIntent hintIntent = FirebaseAuthWrapperFactory
-                .getFirebaseAuthWrapper(mHelper.getAppName())
-                .getEmailHintIntent(getActivity());
-        if (hintIntent != null) {
-            try {
-                startIntentSenderForResult(hintIntent.getIntentSender(), RC_HINT, null, 0, 0, 0, null);
-            } catch (IntentSender.SendIntentException e) {
-                Log.e(TAG, "Unable to start hint intent", e);
-            }
+        try {
+            startIntentSenderForResult(getEmailHintIntent().getIntentSender(), RC_HINT, null, 0, 0, 0, null);
+        } catch (IntentSender.SendIntentException e) {
+            Log.e(TAG, "Unable to start hint intent", e);
         }
+    }
+
+    private PendingIntent getEmailHintIntent() {
+        GoogleApiClient client = new GoogleApiClient.Builder(getContext())
+                .addApi(Auth.CREDENTIALS_API)
+                .enableAutoManage(getActivity(), GoogleApiConstants.AUTO_MANAGE_ID3,
+                                  new GoogleApiClient.OnConnectionFailedListener() {
+                                      @Override
+                                      public void onConnectionFailed(@NonNull ConnectionResult connectionResult) {
+                                          Log.e(TAG,
+                                                "Client connection failed: " + connectionResult.getErrorMessage());
+                                      }
+                                  })
+                .build();
+
+        HintRequest hintRequest = new HintRequest.Builder()
+                .setHintPickerConfig(new CredentialPickerConfig.Builder()
+                                             .setShowCancelButton(true)
+                                             .build())
+                .setEmailAddressIdentifierSupported(true)
+                .build();
+
+        return Auth.CredentialsApi.getHintPickerIntent(client, hintRequest);
     }
 
     @Override
