+++ auth/src/main/java/com/firebase/ui/auth/util/PlayServicesHelper.java
@@ -6,10 +6,14 @@ import android.content.DialogInterface;
 
 import com.google.android.gms.common.GoogleApiAvailability;
 
+/**
+ * Helper class wrapping {@link GoogleApiAvailability}. Used internally but can also be used
+ * by client application
+ */
 public class PlayServicesHelper {
     private static GoogleApiAvailability mApiAvailability;
 
-    private static GoogleApiAvailability getGoogleApiAvailability() {
+    public static GoogleApiAvailability getGoogleApiAvailability() {
         if (mApiAvailability == null) {
             mApiAvailability = GoogleApiAvailability.getInstance();
         }
@@ -17,11 +21,14 @@ public class PlayServicesHelper {
     }
 
     /**
+     * @param activity       The Activity that will host necessary dialogs.
+     * @param requestCode    A request code to be used to return results to the Activity.
+     * @param cancelListener A Dialog listener if the user cancels the recommended action.
      * @return true if play services is available, false otherwise.
      */
     public static boolean makePlayServicesAvailable(Activity activity,
-                                                   int requestCode,
-                                                   DialogInterface.OnCancelListener cancelListener) {
+                                                    int requestCode,
+                                                    DialogInterface.OnCancelListener cancelListener) {
         Dialog errorDialog = getGoogleApiAvailability().getErrorDialog(
                 activity,
                 getGoogleApiAvailability().isGooglePlayServicesAvailable(activity),
