+++ auth/src/test/java/com/firebase/ui/auth/testhelpers/CustomRobolectricGradleTestRunner.java
@@ -19,8 +19,6 @@ import com.firebase.ui.auth.provider.FacebookProvider;
 import com.firebase.ui.auth.provider.GoogleProvider;
 import com.firebase.ui.auth.ui.ActivityHelper;
 import com.firebase.ui.auth.ui.BaseHelper;
-import com.firebase.ui.auth.util.FirebaseAuthWrapperFactory;
-import com.firebase.ui.auth.util.FirebaseAuthWrapperImpl;
 
 import org.junit.runners.model.InitializationError;
 import org.robolectric.RobolectricTestRunner;
@@ -36,10 +34,8 @@ public class CustomRobolectricGradleTestRunner extends RobolectricTestRunner {
     public InstrumentationConfiguration createClassLoaderConfig(Config config) {
         InstrumentationConfiguration.Builder builder = InstrumentationConfiguration.newBuilder();
 
-        builder.addInstrumentedClass(FirebaseAuthWrapperFactory.class.getName());
         builder.addInstrumentedClass(BaseHelper.class.getName());
         builder.addInstrumentedClass(ActivityHelper.class.getName());
-        builder.addInstrumentedClass(FirebaseAuthWrapperImpl.class.getName());
         builder.addInstrumentedClass(FacebookProvider.class.getName());
         builder.addInstrumentedClass(GoogleProvider.class.getName());
         builder.addInstrumentedClass(LoginManager.class.getName());
