+++ auth/src/test/java/com/firebase/ui/auth/ui/email/RegisterEmailActivityTest.java
@@ -15,6 +15,7 @@
 package com.firebase.ui.auth.ui.email;
 
 import android.content.Intent;
+import android.os.Bundle;
 import android.support.design.widget.TextInputLayout;
 import android.widget.Button;
 import android.widget.EditText;
@@ -27,7 +28,6 @@ import com.firebase.ui.auth.testhelpers.AutoCompleteTask;
 import com.firebase.ui.auth.testhelpers.BaseHelperShadow;
 import com.firebase.ui.auth.testhelpers.CustomRobolectricGradleTestRunner;
 import com.firebase.ui.auth.testhelpers.FakeAuthResult;
-import com.firebase.ui.auth.testhelpers.FirebaseAuthWrapperImplShadow;
 import com.firebase.ui.auth.testhelpers.TestConstants;
 import com.firebase.ui.auth.testhelpers.TestHelper;
 import com.firebase.ui.auth.ui.User;
@@ -61,7 +61,7 @@ public class RegisterEmailActivityTest {
                 TestHelper.getFlowParameters(Collections.singletonList(AuthUI.EMAIL_PROVIDER)));
         return Robolectric.buildActivity(RegisterEmailActivity.class)
                 .withIntent(startIntent)
-                .create()
+                .create(new Bundle())
                 .start()
                 .visible()
                 .get();
@@ -73,9 +73,6 @@ public class RegisterEmailActivityTest {
     }
 
     @Test
-    @Config(shadows = {
-            FirebaseAuthWrapperImplShadow.class
-    })
     public void testSignUpButton_validatesFields() {
         RegisterEmailActivity registerEmailActivity = createActivity();
 
@@ -107,8 +104,7 @@ public class RegisterEmailActivityTest {
     @Test
     @Config(shadows = {
             BaseHelperShadow.class,
-            ActivityHelperShadow.class,
-            FirebaseAuthWrapperImplShadow.class
+            ActivityHelperShadow.class
     })
     public void testSignUpButton_successfulRegistrationShouldContinueToSaveCredentials() {
         // init mocks
